requirements.txt
aiohttp==3.5.4
alembic==1.0.10
email-validator==1.0.4
factory-boy==2.12.0
fastapi==0.30.0
psycopg2-binary==2.8.2
pydantic==0.28
PyJWT==1.7.1
pytest==4.5.0
requests==2.22.0
SQLAlchemy==1.3.3
SQLAlchemy-Utils==0.33.11
ujson==1.35
uvicorn==0.8.2
uvloop==0.12.2